wip(tsc): set allowJs in compilerOptions
This is what allows the output literate modules to be written to disk.
Go figure.

Currently, they have an ugly `.ts.md.js` extension, and need to
be imported with a `.ts.md` one. Better extension support in next
version; immediate next step is to try this out in a real use case
in some smart contract client libraries downstream ;-)